zrobione edytowanie i dodawanie postów

+++ tweeter/src/main/java/pl/tweeter/app/entity/Action.java
@@ -10,7 +10,7 @@ public class Action {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "create_timestamp")
     private Timestamp createTimestamp;
@@ -20,18 +20,18 @@ public class Action {
     @Column(name = "expire_timestamp")
     private Timestamp expireTimestamp;
 
-    private Integer userId;
+    private Long userId;
 
-    private Integer postId;
+    private Long postId;
 
     @Column(name = "admin_userId")
-    private Integer adminUserId;
+    private Long adminUserId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -59,27 +59,27 @@ public class Action {
         this.expireTimestamp = expireTimestamp;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(Integer postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
 
-    public Integer getAdminUserId() {
+    public Long getAdminUserId() {
         return adminUserId;
     }
 
-    public void setAdminUserId(Integer adminUserId) {
+    public void setAdminUserId(Long adminUserId) {
         this.adminUserId = adminUserId;
     }
 }
