+++ tweeter/src/main/java/pl/tweeter/app/repository/ActionRepository.java
@@ -3,5 +3,5 @@ package pl.tweeter.app.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.tweeter.app.entity.Action;
 
-public interface ActionRepository extends JpaRepository<Action, Integer> {
+public interface ActionRepository extends JpaRepository<Action, Long> {
 }
