+++ tweeter/src/main/java/pl/tweeter/app/repository/PostRepository.java
@@ -1,10 +1,15 @@
 package pl.tweeter.app.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import pl.tweeter.app.entity.Post;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 
-    Long countByUserId(Integer userID);
+    Long countByUserId(Long userID);
+
+
 
 }
