+++ tweeter/src/main/java/pl/tweeter/app/repository/UserDescriptionRepository.java
@@ -3,5 +3,5 @@ package pl.tweeter.app.repository;
 import org.springframework.data.repository.CrudRepository;
 import pl.tweeter.app.entity.UserDescription;
 
-public interface UserDescriptionRepository extends CrudRepository<UserDescription, Integer> {
+public interface UserDescriptionRepository extends CrudRepository<UserDescription, Long> {
 }
